+++ packages/headless/src/controllers/commerce/field-suggestions/headless-field-suggestions-generator.ts
@@ -5,9 +5,12 @@ import {
   CommerceEngineState,
 } from '../../../app/commerce-engine/commerce-engine';
 import {stateKey} from '../../../app/state-key';
+import {commerceFacetSetReducer as commerceFacetSet} from '../../../features/commerce/facets/facet-set/facet-set-slice';
 import {fieldSuggestionsOrderReducer as fieldSuggestionsOrder} from '../../../features/commerce/facets/field-suggestions-order/field-suggestions-order-slice';
 import {FieldSuggestionsFacet} from '../../../features/commerce/facets/field-suggestions-order/field-suggestions-order-state';
 import {executeSearch} from '../../../features/commerce/search/search-actions';
+import {categoryFacetSearchSetReducer as categoryFacetSearchSet} from '../../../features/facets/facet-search-set/category/category-facet-search-set-slice';
+import {specificFacetSearchSetReducer as facetSearchSet} from '../../../features/facets/facet-search-set/specific/specific-facet-search-set-slice';
 import {loadReducerError} from '../../../utils/errors';
 import {
   buildController,
@@ -65,7 +68,10 @@ export function buildFieldSuggestionsGenerator(
 
   const createFieldSuggestionsControllers = createSelector(
     (state: CommerceEngineState) => state.fieldSuggestionsOrder!,
-    (facetOrder) =>
+    (state: CommerceEngineState) => state.commerceFacetSet,
+    (state: CommerceEngineState) => state.facetSearchSet,
+    (state: CommerceEngineState) => state.categoryFacetSearchSet,
+    (facetOrder, _commerceFacetSet, _facetSearchSet, _categoryFacetSearchSet) =>
       facetOrder.map(({type, facetId}) => {
         switch (type) {
           case 'hierarchical':
@@ -98,6 +104,9 @@ function loadFieldSuggestionsGeneratorReducers(
 ): engine is CommerceEngine {
   engine.addReducers({
     fieldSuggestionsOrder,
+    commerceFacetSet,
+    facetSearchSet,
+    categoryFacetSearchSet,
   });
   return true;
 }