+++ packages/headless/src/features/did-you-mean/did-you-mean-insight-analytics-actions.ts
@@ -6,7 +6,7 @@ import {SearchPageEvents} from '../analytics/search-action-cause';
 import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 
 export const logDidYouMeanClick = (): InsightAction =>
-  makeInsightAnalyticsActionFactory(SearchPageEvents.didyoumeanClick)(
+  makeInsightAnalyticsActionFactory(SearchPageEvents.didYouMeanClick)(
     'analytics/didyoumean/click',
     (client, state) =>
       client.logDidYouMeanClick(
@@ -15,7 +15,7 @@ export const logDidYouMeanClick = (): InsightAction =>
   );
 
 export const logDidYouMeanAutomatic = (): InsightAction =>
-  makeInsightAnalyticsActionFactory(SearchPageEvents.didyoumeanAutomatic)(
+  makeInsightAnalyticsActionFactory(SearchPageEvents.didYouMeanAutomatic)(
     'analytics/didyoumean/automatic',
     (client, state) =>
       client.logDidYouMeanAutomatic(